@@ -1,82 +1,129 @@
 {
   "name": "react-component-stack",
-  "version": "1.0.1",
+  "version": "1.1.2",
   "description": "React Components to build your react project easily.",
   "main": "dist/index.js",
   "module": "dist/index.es.js",
   "jsnext:main": "dist/index.es.js",
+  "private": false,
   "engines": {
     "node": ">=8",
     "npm": ">=5"
   },
   "scripts": {
-    "test": "cross-env CI=1 react-scripts test --env=jsdom",
-    "test:watch": "react-scripts test --env=jsdom",
+    "test": "cross-env CI=1 jest -i --env=jsdom",
+    "tests": "cross-env run-p eslint test",
+    "test:watch": "jest -i --watchAll --env=jsdom",
+    "coverage": "cross-env jest -i --coverage --env=jsdom",
+    "coverage-summary": "cross-env istanbul report --root ./ --include reports/coverage/coverage-final.json --dir ./reports json-summary",
     "build": "rollup -c",
     "start": "rollup -c -w",
     "prepare": "npm run build",
     "predeploy": "cd example && npm install && npm run build",
-    "deploy": "gh-pages -d example/build"
+    "deploy": "gh-pages -d example/build",
+    "eslint": "cross-env eslint 'src/**/*.js' 'src/**/*.jsx'",
+    "reports": "run-p eslint-report coverage-summary",
+    "eslint-json-summary": "npm run eslint -- -f json > ./reports/eslint.json",
+    "eslint-report": "cross-env eslint 'src/**/*.js' 'src/**/*.jsx' -f node_modules/eslint-html-reporter/reporter.js -o reports/eslint.html",
+    "analyze": "source-map-explorer 'dist/*.js'",
+    "storybook": "start-storybook -p 9009",
+    "build-storybook": "build-storybook",
+    "build-storybook:docs": "build-storybook -s public --docs",
+    "chromatic": "chromatic --project-token=wzi80q8s6l"
   },
   "peerDependencies": {
-    "@material-ui/core": "^3.9.3",
-    "@material-ui/icons": "^3.0.1",
-    "@material-ui/lab": "^3.0.0-alpha.25",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "16.4.0",
-    "react-device-detect": "^1.6.2"
+    "@material-ui/core": "^4.9.8",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.47",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "dependencies": {
     "@date-io/moment": "^1.3.5",
-    "@tinymce/tinymce-react": "^3.2.0",
-    "async": "^2.6.2",
-    "autosuggest-highlight": "^3.1.1",
+    "@material-ui/pickers": "^3.2.9",
     "classnames": "2.2.5",
-    "lodash": "^4.17.11",
-    "material-ui-pickers": "^2.2.4",
-    "moment": "^2.24.0",
-    "moment-range": "^4.0.2",
-    "react-autosuggest": "^9.4.2",
-    "react-daterange-picker": "^2.0.1",
-    "react-router-dom": "4.2.2",
-    "react-select": "^2.4.3",
-    "tinymce": "^5.0.8"
+    "flat": "^5.0.0",
+    "prop-types": "^15.7.2",
+    "validator": "^13.1.1"
   },
   "devDependencies": {
+    "@babel/plugin-external-helpers": "^7.8.3",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "@material-ui/core": "^4.9.8",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.47",
+    "@storybook/addon-actions": "^5.3.19",
+    "@storybook/addon-docs": "^5.3.19",
+    "@storybook/addon-links": "^5.3.19",
+    "@storybook/addons": "^5.3.19",
+    "@storybook/react": "^5.3.19",
     "@svgr/rollup": "^2.4.1",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.5",
-    "babel-plugin-external-helpers": "^6.22.0",
+    "autoprefixer": "^9.8.0",
+    "babel-core": "^7.0.0-beta.41",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "^8.1.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
+    "chromatic": "^4.0.3",
     "cross-env": "^5.1.4",
-    "eslint": "^5.0.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
+    "enzyme-to-json": "^3.4.4",
+    "eslint": "^6.7.0",
     "eslint-config-standard": "^11.0.0",
     "eslint-config-standard-react": "^6.0.0",
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-promise": "^4.0.0",
     "eslint-plugin-react": "^7.10.0",
+    "eslint-plugin-react-hooks": "^3.0.0",
     "eslint-plugin-standard": "^3.1.0",
     "fs": "0.0.1-security",
     "gh-pages": "^1.2.0",
+    "husky": "^4.2.5",
+    "jest": "^24.9.0",
+    "lint-staged": "^10.2.6",
+    "moment": "^2.26.0",
     "path": "^0.12.7",
-    "react": "^16.4.1",
+    "prettier": "^2.0.5",
+    "react": "^16.12.0",
     "react-docgen": "^4.1.1",
-    "react-dom": "^16.4.1",
-    "react-scripts": "^1.1.4",
-    "rollup": "^0.64.1",
-    "rollup-plugin-babel": "^3.0.7",
+    "react-dom": "^16.12.0",
+    "react-test-renderer": "^16.13.1",
+    "rollup": "^2.3.0",
+    "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-commonjs": "^9.1.3",
+    "rollup-plugin-css-bundle": "^1.0.4",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-peer-deps-external": "^2.2.0",
     "rollup-plugin-postcss": "^1.6.2",
-    "rollup-plugin-url": "^1.4.0"
+    "rollup-plugin-url": "^1.4.0",
+    "source-map-explorer": "^2.4.2"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
+      "prettier --write"
+    ]
   },
   "files": [
-    "dist"
+    "src/**/*.js",
+    "!src/**/*.stories.js",
+    "!src/**/*.spec.js",
+    "dist",
+    "README.md"
+  ],
+  "pre-commit": [
+    "tests"
   ],
   "repository": {
     "type": "git",
