@@ -1,13 +1,1 @@
-import Range from "./Range";
-import Select from "./Select";
-import Toggle from "./Toggle";
-import DateRange from "./DateRange";
-import AutoComplete from "./AutoComplete";
-import AutoSuggest from "./AutoSuggest";
-import RichText from "./RichText";
-import Number from "./Number";
-import Radio from "./Radio";
-
-export { 
-	AutoComplete, AutoSuggest, Range, Select, Toggle, DateRange, RichText, Number, Radio 
-};
+export { default } from "./Field";
