@@ -1,19 +1,3 @@
-export { default as Accordion } from './Accordion';
-export { default as Card } from './Card';
-export { default as ChipFilter } from './ChipFilter';
-export { default as Dialog } from './Dialog';
-export { default as Drawer } from './Drawer';
-export { default as Field } from './Field';
-export { default as Filter } from './Filter';
-export { default as Form } from './Form';
-export { default as Footer } from './Footer';
-export { default as GridList } from './GridList';
-export { default as Table } from './Table/Table';
-export { default as List } from './List';
-export { default as Popper } from './Popper';
-export { default as Slider } from './Slider';
-export { default as StatusBar } from './StatusBar';
-export { default as Stepper } from './Stepper';
-export { default as Snackbar } from './Snackbar';
-export { default as Sort } from './Sort';
-export { default as Toolbar } from './Toolbar';
+export { default as Form } from "./Form";
+export { default as Table } from "./Table";
+export { default as FormField } from "./FormField";
